更新
1. Auth 服务模块
2. Captcha 服务模块
3. License 服务模块
4. Mail 服务模块
5. SignIn 服务模块
6. Users 服务模块

+++ xuyou-user/src/main/java/ltd/xiaomizha/xuyou/user/entity/UserResourceRelations.java
@@ -82,7 +82,7 @@ public class UserResourceRelations implements Serializable {
     /**
      * 授权人用户ID
      */
-    private Integer grantedBy;
+    private Long grantedBy;
 
     /**
      * 授权原因
@@ -107,7 +107,7 @@ public class UserResourceRelations implements Serializable {
     /**
      * 撤销人用户ID
      */
-    private Integer revokedBy;
+    private Long revokedBy;
 
     /**
      * 撤销原因
